Refactor library and engine structure
Refactors library and engine structure to be less coupled. Pulls classes
and interfaces out into individual files. Additionally, renames package
from util to engine.
